add Camera utils and move isFrontCameraAvailable() there
